+++ src/backend/ci/core/process/api-process/src/main/kotlin/com/tencent/devops/process/pojo/template/v2/PipelineTemplateInfo.kt
@@ -26,7 +26,7 @@ data class PipelineTemplateInfo(
     @get:Schema(title = "是否开启PAC", required = true)
     val enablePac: Boolean,
     @get:Schema(title = "最新发布版本号", required = true)
-    val releasedVersion: Int? = null,
+    val releasedVersion: Long? = null,
     @get:Schema(title = "最新发布版本名称", required = false)
     val releasedVersionName: String? = null,
     @get:Schema(title = "最新发布配置版本号", required = false)
